% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_origin_axes.R
\name{add_origin_axes}
\alias{add_origin_axes}
\title{Draw colored axis lines}
\usage{
add_origin_axes(
  xmin = -80,
  xmax = 80,
  ymin = -80,
  ymax = 80,
  zmin = -80,
  zmax = 80,
  xcol = "red",
  ycol = "darkgreen",
  zcol = "darkblue",
  lwd = 5
)
}
\arguments{
\item{xmin, xmax}{minimum and maximum extents of the lines in x direction}

\item{ymin, ymax}{minimum and maximum extents of the lines in y direction}

\item{zmin, zmax}{minimum and maximum extents of the lines in z direction}

\item{xcol, ycol, zcol}{colors for each line; specify with a quoted color name
or hex code}

\item{lwd}{width of lines (see \code{\link[rgl:material]{rgl::rgl.material()}})}
}
\value{
No return value; function is called for its side-effect of rendering
the axes in the existing viewer window.
}
\description{
\strong{rgl} graphics device does not support axis lines such as those seen in
\emph{MeshLab}. This function makes it simple to verify where the origin (i.e.
0,0,0) lies in the viewer.
}
