% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_xy.R
\name{adjust_xy}
\alias{adjust_xy}
\title{Correct x and y offset for a soil specimen}
\usage{
adjust_xy(mesh)
}
\arguments{
\item{mesh}{a "mesh3d" object which is already oriented with its surface lying parallel to the x-y plane}
}
\value{
a "mesh3d" object with its center lying in the normal to the x-y plane
}
\description{
After re-orienting and moving the specimen with \link{orient_sample}, the minute error in the horizontal plane needs to be corrected prior to removing the aluminum sample holder and the outermost portion of the soil surface. This function computes the difference between the maximum and minimum values observed in the x and y directions and corrects the vertex coordinates so the center is truly located at x=0 and y=0.
}
\examples{
adjusted <- adjust_xy(mesh = untrimmed_mesh1)
rgl::shade3d(untrimmed_mesh1, color="firebrick")
rgl::shade3d(adjusted, color= "darkblue")
add_origin_axes()
}
