% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_z.R
\name{adjust_z}
\alias{adjust_z}
\title{Shift a mesh up or down}
\usage{
adjust_z(mesh, n_bins = 10000)
}
\arguments{
\item{mesh}{a "mesh3d" object}

\item{n_bins}{number of bins among which the coordinates should be split}
}
\value{
a "mesh3d" object with adjusted vertex coordinates
}
\description{
After using \code{\link[=adjust_xy]{adjust_xy()}} to center the mesh, it must be moved up or down so
its origin lies in the x-y plane. \code{adjust_z()} computes the 10 most frequent
values in the mesh's z-coordinates and translates the mesh by this amount,
accounting for the direction of the "offset."
}
\examples{
xy_fixed <- remove_cyl(adjust_xy(untrimmed_mesh1))
z_fixed <- remove_cyl(adjust_z(xy_fixed))
rgl::shade3d(xy_fixed, alpha=1/2)
rgl::shade3d(z_fixed)
add_origin_axes()
}
