% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_lights.R
\name{better_lights}
\alias{better_lights}
\title{Wrapper to make rgl lighting parameters less shiny}
\usage{
better_lights()
}
\value{
alters existing rgl scene or opens a new one
}
\description{
Wrapper to make rgl lighting parameters less shiny
}
