% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_cyl_dims.R
\docType{data}
\name{chunk_cyl_dims}
\alias{chunk_cyl_dims}
\title{Masses and volumes of specimen molds for cleat-mark device}
\format{
A tibble with 3 columns and 12 rows (one for each cylinder)
}
\usage{
chunk_cyl_dims
}
\description{
Twelve aluminum cylinders were custom fabricated in February
2020. These are 6 1/16" in diameter and 4" deep with a cylindrical bottom
and a small lip for fastening into the device. For quality control of
compaction measurements, the volume and mass of each cylinder must be known
precisely.

Each cylinder also was retro-fitted with a low-density resin plug
approximately 1" thick. The need for this modification arose from the fact
that the mass of a full cylinder (i.e. with 4" of compacted moist soil
inside) exceeds the rated capacity of the 3D scanning turntable. The resin
fills a portion of the cylinder volume that would otherwise be occupied by
soil, reducing the total specimen mass to < 5 kg.

The dimensions of each cylinder and resin plug were measured on 2020-09-10.
The process was similar to that outlined in \href{https://www.astm.org/Standards/D698.htm}{ASTM D 698-12e, Appendix A}. De-aired water was used to
fill the cylinders and excess water extending above the surface via surface
tension was carefully removed by lowering a glass plate onto the surface
and catching any water which spilled down the exterior with a cloth towel.
This process was performed \emph{without} the resin plug inside the cylinder, as
its low bulk density encouraged the plug to float and prevented accurate
water measurements from being obtained.

Three replications were performed for each cylinder and the values were
averaged to yield one volume per cylinder.

The resin volume was computed by measuring the plug's thickness to ± 0.01
mm in 3 locations with a digital caliper. The volume was computed from this
thickness and the known cylinder diameter of 6 1/16". The resin volume was
subtracted from the water-measured volume to give the actual empty volume
available to compacted soil.
}
\keyword{datasets}
