% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleat_mark_backfill_datasheet.R
\name{cleat_mark_backfill_datasheet}
\alias{cleat_mark_backfill_datasheet}
\title{Populate empty data sheet for filling in cleats with sand}
\usage{
cleat_mark_backfill_datasheet(
  experiment_name,
  date,
  tin_tare_set,
  sand_loose_density = asi468::vrc_sand_loose_density
)
}
\arguments{
\item{experiment_name}{Character string identifying the experiment}

\item{date}{Date of data collection (YYYY-MM-DD)}

\item{tin_tare_set}{Character string identifying the date corresponding to
the tin tares in \strong{asi468}}

\item{sand_loose_density}{Bulk density of sand used for sand-cone test,
defaults to \code{asi468::vrc_sand_loose_density}, which is a data object stored
in the asi468 package. Its value is 1.54, as was measured in Dec. 2020 for
the topdressing sand obtained from Valentine research center.}
}
\value{
writes file to disk
}
\description{
Populate empty data sheet for filling in cleats with sand
}
