% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleat_mark_prep_datasheet.R
\name{cleat_mark_prep_datasheet}
\alias{cleat_mark_prep_datasheet}
\title{Populate a Blank csv File for Quality Control}
\usage{
cleat_mark_prep_datasheet(dir, date, tin_tare_set = "")
}
\arguments{
\item{dir}{Character. Directory to which file should be written.}

\item{date}{Character. Date specimens were prepared; ISO format (YYYY-MM-DD)}

\item{tin_tare_set}{Character. Identifier corresponding to the set of tins used to measure water content. See \link[asi468:tin_tares]{asi468::tin_tares}.}
}
\value{
Writes file to disk and prints a message.
}
\description{
Simplifies the process of measuring dry density and water content
of compacted specimens. Dry density should correspond closely to
that achieved with Proctor testing. Column names are mostly compatible
with \code{\link[soiltestr:add_physical_properties]{soiltestr::add_physical_properties()}}; the exception is that
the properties of the cylinders themselves (stored in \code{\link{chunk_cyl_dims}} have different names to make it clear that their
masses and volumes include the foam plugs). These are automatically
converted by \code{\link[soiltestr:add_physical_properties]{soiltestr::add_physical_properties()}}.
}
