% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_cross_sect_area_cm2.R
\docType{data}
\name{cyl_cross_sect_area_cm2}
\alias{cyl_cross_sect_area_cm2}
\title{Cross-sectional area of one cylinder}
\format{
numeric vector of length 1'
}
\usage{
cyl_cross_sect_area_cm2
}
\description{
The pipe is 6 1/16" inner diameter. I often use the
cross-sectional area for calculations, so I am storing it here as a data
object in the \strong{soilmesh} package.
}
\keyword{datasets}
