% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirtGIF3d.R
\name{dirtGIF3d}
\alias{dirtGIF3d}
\title{Save a GIF image of a spinning soil surface}
\usage{
dirtGIF3d(
  mesh,
  output_file,
  phi = -35,
  zoom = 0.8,
  n_snaps = 10,
  fps = 2,
  bg_color = "grey60"
)
}
\arguments{
\item{mesh}{mesh3d object to record}

\item{output_file}{relative file path to save (quoted, ending in ".gif")}

\item{phi}{vertical observer angle, see \link[rgl]{rgl.viewpoint}}

\item{zoom}{zooming factor, larger number is smaller object. Defaults to 0.8.}

\item{n_snaps}{number of snapshots to concatenate}

\item{fps}{speed at which frames are displayed (frames per second)}

\item{bg_color}{quoted color name or hex code for the scene background}
}
\value{
no return value; file is saved to disk
}
\description{
Showing the 3D scans in motion is a useful way to visualize cleat indentations
and visualize the surface microtopography.
}
