% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drydown_analysis.R
\name{drydown_analysis}
\alias{drydown_analysis}
\title{Quality control for drying soil specimens}
\usage{
drydown_analysis(path)
}
\arguments{
\item{path}{path to raw data file}
}
\value{
3-item list of class \code{"soilmesh_drydown"}. All water content references are on a gravimetric basis
\enumerate{
\item \code{water_losses} Tibble containing drying time, total water loss, and rate of water loss per hour
\item \code{water_loss_summary} Tibble containing average water loss data across all 3 cylinders of the same soil
\item \code{drydown_plot}ggplot showing drydown information for each soil
}
}
\description{
Take an input data file and calculate water loss for each
specimen, then plot the results. A linear rate of water loss must be
assumed because only 2 data points are collected; however strictly speaking
the water loss usually is linear above the shrinkage limit and then becomes
curvilinear as the soil dries further.
}
\examples{
# # create temporary csv file from built-in dataset
#
# f <- file.path(tempdir(), "example_drydown_file.csv")
# readr::write_csv(x = soilmesh::example_drydown_data, file = f)
#
# data <- readr::read_csv(f)
#
#
# # perform analysis
#
# eda <- drydown_analysis(path = f)
}
