% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drydown_datasheet.R
\name{drydown_datasheet}
\alias{drydown_datasheet}
\title{Populate a skeleton csv file for data collection with cleat-mark
device}
\usage{
drydown_datasheet(soil_IDs, date, dir)
}
\arguments{
\item{soil_IDs}{character vector of the unique identifiers of the soils being
tested (length 4)}

\item{date}{date of data collection ("yyyy-mm-dd")}

\item{dir}{directory in which to save the .csv file}
}
\value{
File is written to disk and a message is printed in the console.
}
\description{
Ensures consistent column names and rectangular format across
days and soils. Records information needed to calculate water content at two sampling times.
}
\details{
This format allows a water content sample to be extracted at the
beginning of the dry-down process and a second sample to be extracted
immediately after testing. This allows the total water loss to be computed
over the course of the test. Although the rate of water loss may vary as a
function of water content, this is still a useful data point for estimating
evaporation rate, and also because it allows a check to be performed on the
uniformity of water content across cylinders of the same soil at the
beginning of the test. See \code{\link[=drydown_analysis]{drydown_analysis()}}.
}
