% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_drydown_data.R
\docType{data}
\name{example_drydown_data}
\alias{example_drydown_data}
\title{Example data object containing data from one day of specimen drying}
\format{
A tibble
}
\usage{
example_drydown_data
}
\description{
Example data object containing data from one day of specimen drying
}
\keyword{datasets}
