% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_topo_colors.R
\name{mesh_topo_colors}
\alias{mesh_topo_colors}
\title{Show the z-elevations of a mesh with color}
\usage{
mesh_topo_colors(mesh, file = NULL, palette = NULL, origin_axes = FALSE, ...)
}
\arguments{
\item{mesh}{mesh3d object}

\item{file}{optional, provide a file stem (no extension) if the plot and mesh are to be saved}

\item{origin_axes}{plot the axes?}

\item{...}{other arguments passed to \code{Morpho::meshDist()}}
}
\value{
plots the colored mesh in an rgl window
}
\description{
Show the z-elevations of a mesh with color
}
