% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_cleat_mark_datasheets.R
\name{new_cleat_mark_datasheets}
\alias{new_cleat_mark_datasheets}
\title{Create file structure and empty data sheets for cleat mark testing}
\usage{
new_cleat_mark_datasheets(date, experiment_name, sample_names)
}
\arguments{
\item{date}{data collection date ("YYYY-MM-DD")}

\item{experiment_name}{character string uniquely identifying this set of soils or mixtures}

\item{sample_names}{character vector of the unique identifiers of the soils being
tested (length 4)}
}
\value{
writes new folders and empty data files to disk
}
\description{
Creates 2 directories and 2 additional data sheets. Designed to
store penetrometer data, color photo indexes, and raw mesh files.
}
