% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_cleatmark_report.R
\name{new_cleatmark_report}
\alias{new_cleatmark_report}
\title{Create a new report based on a template}
\usage{
new_cleatmark_report(file, ...)
}
\arguments{
\item{file}{output path for report .Rmd file}

\item{...}{other arguments passed to \code{\link[rmarkdown:draft]{rmarkdown::draft()}}}
}
\value{
opens a new file in the RStudio IDE
}
\description{
Report contains plots of mesh metrics and images of the 3D mesh files.
YAML is populated to support both HTML and PDF output.
}
