% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_mini_density_datasheet.R
\name{new_mini_density_datasheet}
\alias{new_mini_density_datasheet}
\title{Populate empty data sheet for miniature sand-cone test}
\usage{
new_mini_density_datasheet(
  experiment_ID,
  date,
  tin_tare_set,
  sand_loose_density = 1.54
)
}
\arguments{
\item{experiment_ID}{Character string identifying the experiment}

\item{date}{Date of data collection (YYYY-MM-DD)}

\item{tin_tare_set}{Character string identifying the date corresponding to
the tin tares in \strong{asi468}}

\item{sand_loose_density}{Bulk density of sand used for sand-cone test,
defaults to 1.54 as was measured in Dec. 2020 for the topdressing sand
obtained from Valentine research center.}
}
\value{
silently writes file to disk
}
\description{
A helper function wrapped by \code{\link[=new_cleat_mark_datasheets]{new_cleat_mark_datasheets()}}
}
