% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overhead-snapshot-fns.R
\name{overhead_mesh_snapshot}
\alias{overhead_mesh_snapshot}
\title{Save an overhead snapshot of a mesh and trim the white space}
\usage{
overhead_mesh_snapshot(mesh = NULL, file, mesh_path = NULL, ...)
}
\arguments{
\item{mesh}{mesh3d object to screenshot}

\item{file}{path to save the screenshot}

\item{mesh_path}{optionally allows mesh to be snapshotted directly from a file path. If \code{mesh} is provided, this argument is ignored}

\item{...}{arguments passed to \code{rgl::shade3d()}}
}
\value{
Writes file to disk
}
\description{
Save an overhead snapshot of a mesh and trim the white space
}
