% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overhead-snapshot-fns.R
\name{overhead_mesh_view}
\alias{overhead_mesh_view}
\title{Show an overhead view of a mesh, with a square viewport and better lights}
\usage{
overhead_mesh_view()
}
\value{
alters existing rgl scene
}
\description{
Show an overhead view of a mesh, with a square viewport and better lights
}
