% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mesh_filename.R
\name{parse_mesh_filename}
\alias{parse_mesh_filename}
\title{Split the base file name of a mesh scan into atomic idenifiers}
\usage{
parse_mesh_filename(x)
}
\arguments{
\item{x}{Character vector of \code{.ply} files to read}
}
\value{
Data frame containing 5 columns; the first holds the full file path
while the rest hold the data which uniquely identify the file (See details)
}
\description{
A file naming sheme was developed for this project to ensure all
data files are uniquely identified. Each \code{.ply} file contains 4 pieces of
information (see details). This function parses the file name into useful
metadata about the specimen. Returns a tibble for use subsequent analysis
with \strong{tidyverse} tools.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Each file name contains 4 pieces of information, separated by
underscores. These provide enough information to uniquely identify every
scan, and the additional information about the soil can be added via
joining other data frames. The pieces of information encoded in the file
names are (in order of appearance):
\enumerate{
\item \code{experiment_ID}: A two-digit identifier or a character string
corresponding to the particular set of mixes or task. 2. \code{soil_ID} 3-digit
number or other character string identifying the soil \emph{within} the
experiment 3. \code{date} Calendar date of the scan in POSIX format, i.e.
YYYY-MM-DD 4. \code{cylinder_ID} The aluminum sample holder containing the
specimen of interest
}
}
