% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_circ.R
\docType{data}
\name{ref_circ}
\alias{ref_circ}
\title{A round mesh lying in the x-y plane, with radius of 70 mm}
\format{
An object of class \code{mesh3d} of length 3.
}
\usage{
ref_circ
}
\description{
This "mesh3d" object is meant for "glueing" to all data points lying either
above or below the x-y plane. This action facilitates a volume calculation
for disturbed soil, via the functions \code{vol_below()} and \code{vol_above()}.
}
\keyword{datasets}
