% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_cyl.R
\name{remove_cyl}
\alias{remove_cyl}
\title{\lifecycle{experimental}}
\usage{
remove_cyl(mesh, radius = 70, clip_fun = "default", keep = "inside", ...)
}
\arguments{
\item{mesh}{The "mesh3d" object to trim}

\item{radius}{Radius of clipping surface in mm}

\item{clip_fun}{A function corresponding to the bounding surface, defaults to an open cylinder}

\item{keep}{"inside" for vertices contained by the surface; "outside" for those falling beyond it}

\item{...}{allows user to set the \code{attribute} argument if needed (see \link[rgl]{clipMesh3d})}
}
\value{
A "mesh3d" object with undesired vertices removed
}
\description{
Remove the specimen container from an oriented 3D mesh
}
\details{
Boolean mesh operation to keep all points inside or outside the clipping surface
}
\examples{
untrimmed <- rgl::translate3d(untrimmed_mesh1, x=-160, y=0, z=0)
trimmed <- remove_cyl(untrimmed_mesh1)
rgl::shade3d(untrimmed)
rgl::shade3d(trimmed)
}
