% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl_params_16x9.R
\docType{data}
\name{rgl_params_16x9}
\alias{rgl_params_16x9}
\title{Viewer parameters for a 16x9 screenshot with the rgl window}
\format{
a list
}
\usage{
rgl_params_16x9
}
\description{
Still not quite correct but at least I know how to save them
}
\keyword{datasets}
