% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl_params_4x3.R
\docType{data}
\name{rgl_params_4x3}
\alias{rgl_params_4x3}
\title{Viewer parameters for a 4x3 screenshot with the rgl window}
\format{
a list
}
\usage{
rgl_params_4x3
}
\description{
Still not quite correct but at least I know how to save them
}
\keyword{datasets}
