% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmatrices.R
\docType{data}
\name{rtmatrices}
\alias{rtmatrices}
\title{List of roto-translation matrices manually calibrated with MeshLab}
\format{
a list of data frames (each having the "tbl_df" class) with names
corresponding to the date they were collected. The sole matrix from 2020 is
the method development matrix and should not be used for later analyses.
}
\usage{
rtmatrices
}
\description{
This list object allows the scanner to be moved around and
re-calibrated without losing the ability to easily process the meshes with
a rigid rotation and translation to return the mesh to real-world
coordinates.
}
\keyword{datasets}
