% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untrimmed_mesh1.R
\docType{data}
\name{untrimmed_mesh1}
\alias{untrimmed_mesh1}
\title{A soil specimen in need of processing}
\format{
A .ply file with ~130,000 faces
}
\source{
scan produced by package author
}
\usage{
untrimmed_mesh1
}
\description{
This 3D soil specimen has nicely-shaped cleat marks.
}
\keyword{datasets}
