% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh-volume-fns-april-2021.R
\name{vol_below}
\alias{vol_below}
\title{Mesh volume below datum}
\usage{
vol_below(mesh, z_displ_mm = 0.01)
}
\arguments{
\item{mesh}{triangular mesh of class "mesh3d"}

\item{z_displ_mm}{vertical offset to prevent removal of vertices near the origin}
}
\description{
Mesh volume below datum
}
