% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh-volume-fns-april-2021.R
\name{vol_diff}
\alias{vol_diff}
\title{Calculate above, below, and summation of these}
\usage{
vol_diff(mesh, z_displ_mm = 0.01)
}
\arguments{
\item{mesh}{triangular mesh of class "mesh3d"}

\item{z_displ_mm}{vertical offset to prevent removal of vertices near the origin}
}
\value{
named list of length 3
}
\description{
Returns a list with each named element being a numeric vector
of length 1
}
