% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_faceweighted.R
\name{vol_faceweighted}
\alias{vol_faceweighted}
\title{Compute volume departures from reference plane}
\usage{
vol_faceweighted(mesh, radius = 7)
}
\arguments{
\item{mesh}{Object of class "mesh3d"}

\item{radius}{Radius in cm that the mesh was trimmed, see
\code{\link[=pre_process_mesh]{pre_process_mesh()}} and \code{\link[=remove_cyl]{remove_cyl()}}}
}
\description{
Computes area of each face and average distance of its 3
vertices. Multiplies these values to yield a weighted "approximate volume"
for each face. Sums the values based on whether the mean distance is above
or below the reference # datum.
}
\examples{
library(magrittr)

trimmed <- remove_cyl(untrimmed_mesh1) \%>\%
  adjust_xy() \%>\%
  adjust_z()


rgl::clear3d()
mesheR::wiremesh3d(trimmed)
add_origin_axes()
mesheR::wiremesh3d(soilmesh::ref_circ)

vols <- vol_faceweighted(trimmed, radius = 7)

}
