% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh-volume-fns.R
\name{vol_gridded}
\alias{vol_gridded}
\title{\lifecycle{experimental}
Compute volume difference based on signed distance field}
\usage{
vol_gridded(mesh)
}
\arguments{
\item{mesh}{object of class "mesh3d"}
}
\value{
list of length 3 (above, below, total volumes in cm^3^)
}
\description{
Uses number of points above and below datum to compute a weighted
average of the soil volume above and below datum
}
