% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh-volume-fns.R
\name{vol_slice_above}
\alias{vol_slice_above}
\title{\lifecycle{experimental}
Mesh volume above datum}
\usage{
vol_slice_above(mesh, z_displ_mm = 0.01)
}
\arguments{
\item{mesh}{triangular mesh of class "mesh3d"}

\item{z_displ_mm}{vertical offset to prevent removal of vertices near the origin}
}
\description{
Returns sliced mesh and volume computed by re-meshing method
}
