% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_total.R
\name{vol_total}
\alias{vol_total}
\title{Compute total volume of disturbed soil after a cleat-mark test}
\usage{
vol_total(mesh)
}
\arguments{
\item{mesh}{triangular mesh of class "mesh3d"}
}
\value{
volume of soil disturbed during the test in cm3
}
\description{
Makes internal calls to \code{\link[=vol_below]{vol_below()}} and \code{\link[=vol_above]{vol_above()}} and
sums the results.
}
\details{
\lifecycle{experimental}
}
