% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process_meshes.R
\name{batch_process_meshes}
\alias{batch_process_meshes}
\title{Re-orient and trim meshes, then re-write to new files}
\usage{
batch_process_meshes(
  date,
  rtmatrix = NULL,
  downsample = TRUE,
  n_faces = 250000,
  recolor = TRUE,
  write_cols = TRUE,
  save_metrics = TRUE,
  dne_tarface = 10000,
  snapshots = TRUE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{date}{character string corresponding to date 3D scans were collected}

\item{rtmatrix}{a 4x4 matrix. Normally supplied by passing a character string
corresponding to one of the YYYY-MM-DD dates in the \code{soiltestr::rtmatrices}
data object. Alternatively, pass a matrix either from the aforementioned
data object or a matrix present as a variable in the existing global
environment. \code{\link[=rtmatrices]{rtmatrices()}}}

\item{downsample}{Whether to reduce the number of faces by calling
\code{\link[Rvcg:vcgQEdecim]{Rvcg::vcgQEdecim()}}}

\item{n_faces}{if \code{downsample = TRUE}, number of faces to which mesh should
be reduced. Defaults to 250k. Over-ride to a percent by supplying a numeric
value with \verb{percent = <0-1>}, which is passed to  \code{\link[Rvcg:vcgQEdecim]{Rvcg::vcgQEdecim()}}}

\item{recolor}{whether to re-color the meshes based on the average hex color of a digital image taken during testing}

\item{write_cols}{whether to write a csv file containing the hex color codes for each specimen (into the /derived_data/\<date\> directory)}

\item{save_metrics}{whether to write a csv file containing morphometrics data derived from this set of meshes}

\item{dne_tarface}{number of faces to further downsample mesh to compute DNE metric}

\item{snapshots}{whether to capture screenshots of the meshes (by default, both 16x9 and 4x3 are written)}

\item{overwrite}{whether to over-write existing files in the target directory
(if they have identical names)}

\item{...}{Other arguments passed to \code{\link[Rvcg:vcgQEdecim]{Rvcg::vcgQEdecim()}}}
}
\value{
Writes files to disk and prints a message.
}
\description{
This function automates a process that I do repeatedly after
collecting 3D scans.
}
\details{
\lifecycle{experimental}
}
