% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process_meshes.R
\name{batch_process_meshes}
\alias{batch_process_meshes}
\title{Re-orient and trim meshes, then re-write to new files}
\usage{
batch_process_meshes(
  input_directory,
  output_directory,
  rtmatrix = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{input_directory}{path to a directory containing one or more \code{.ply} files}

\item{output_directory}{path to a directory where the new files will be written}

\item{rtmatrix}{character string corresponding to one of the 4x4 matrices from the elements of \code{\link[=rtmatrices]{rtmatrices()}}}

\item{overwrite}{whether to over-write existing files in the target directory (if they have identical names)}
}
\value{
Writes files to disk and prints a message
}
\description{
This function automates a process that I do repeatedly after collecting 3D scans.
}
