% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleat_mark_backfill_datasheet.R
\name{cleat_mark_backfill_datasheet}
\alias{cleat_mark_backfill_datasheet}
\title{Populate empty data sheet for filling in cleats with sand}
\usage{
cleat_mark_backfill_datasheet(
  experiment_name,
  date,
  tin_tare_set,
  bowl_tare_set,
  sand_loose_density = asi468::vrc_sand_loose_density
)
}
\arguments{
\item{experiment_name}{character string uniquely identifying this set of
soils or mixtures}

\item{date}{data collection date ("YYYY-MM-DD")}

\item{tin_tare_set}{unique ID for the set of tins used to measure water
content; see \link[asi468:tin_tares]{asi468::tin_tares}}

\item{bowl_tare_set}{unique ID for the set of tins used to measure water
content; see \link[asi468:tin_tares]{asi468::tin_tares}}

\item{sand_loose_density}{Value for loose dry density of material used to
fill holes. Default value stored as \link[asi468:vrc_sand_loose_density]{asi468::vrc_sand_loose_density}, which
is 1.54.}
}
\value{
writes file to disk
}
\description{
Populate empty data sheet for filling in cleats with sand
}
