% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleat_mark_datasheets.R
\name{cleat_mark_datasheets}
\alias{cleat_mark_datasheets}
\title{Create file structure and empty data sheets for cleat mark testing}
\usage{
cleat_mark_datasheets(
  experiment_name,
  sample_names,
  date,
  tin_tare_set = NULL,
  bowl_tare_set = NULL,
  sand_loose_density = asi468::vrc_sand_loose_density,
  mini_density_reps = 1,
  drydown_tin_numbers = NULL,
  mini_density_tin_numbers = NULL
)
}
\arguments{
\item{experiment_name}{character string uniquely identifying this set of
soils or mixtures}

\item{sample_names}{character vector of the unique identifiers of the soils
being tested (length 4)}

\item{date}{data collection date ("YYYY-MM-DD")}

\item{tin_tare_set}{Optional character string (length 1) corresponding to set of tin tares used for weighing soil}

\item{mini_density_reps}{number of replicate specimens extracted per cylinder
for the miniature sand-cone method}

\item{drydown_tin_numbers}{Optional character vector corresponding (length 24) to specific tin tare numbers used for weighing soil in drydown comparison}

\item{mini_density_tin_numbers}{Optional character vector corresponding (length 12) to specific tin tare numbers used for weighing soil in mini-density measurements}
}
\value{
Writes new folders and empty data files to disk.
}
\description{
Creates 2 directories and 2 additional data sheets. Designed to
store penetrometer data, color photo indexes, and raw mesh files. A new
directory is created for the date of testing, and this holds all files
related to cleat-mark testing for that day. are created relative to root or
current working directory. The by-date directories should live under
\verb{ecmdata/raw-data/cleat-mark-testing/} for a given project.
}
