% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meshfiles.R
\name{read_meshfiles}
\alias{read_meshfiles}
\title{Read multiple 3D mesh files}
\usage{
read_meshfiles(dir = NULL, ...)
}
\arguments{
\item{dir}{Directory in which to search for \code{.ply} files}

\item{...}{additional arguments passed to \code{\link[Rvcg:vcgImport]{Rvcg::vcgImport()}}}
}
\description{
A specific function to automate reading of \code{.ply} files into R. Specify a
directory containing the files to automatically read all contained \code{.ply}
files into R and compile them into a nested tibble with one row per file.
Result can be piped to \code{\link[=parse_mesh_filename]{parse_mesh_filename()}} prior to mesh
processing/analysis.
}
\details{
Wraps \code{\link[=list.files]{list.files()}} from base R, \code{\link[stringr:str_remove]{stringr::str_remove()}},
\code{\link[purrr:map]{purrr::map()}}, and \code{\link[Rvcg:vcgImport]{Rvcg::vcgImport()}} with specific arguments to
locate files, import to current session , strip file extensions, and compile
in tidy format.
}
\seealso{
\code{\link[ecmfuns:gather.files]{ecmfuns::gather.files()}}, \code{\link[ecmfuns:read.files]{ecmfuns::read.files()}}, \code{\link[Rvcg:vcgImport]{Rvcg::vcgImport()}}
}
