% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_above.R
\name{vol_above}
\alias{vol_above}
\title{Compute disturbed soil volume \strong{above} original surface plane}
\usage{
vol_above(mesh, show = TRUE)
}
\arguments{
\item{mesh}{triangular mesh of class "mesh3d"}

\item{show}{whether to display the sliced mesh in a new \strong{rgl} window.}
}
\value{
mesh volume in mm3
}
\description{
Slice the mesh at the z=0 plane and keep only points at or above
the plane. Then merge it with the reference plane of the same diameter.
Compute the volume in mm3.
}
\details{
\lifecycle{experimental}
}
