% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh-volume-fns.R
\name{vol_faceweighted}
\alias{vol_faceweighted}
\title{Compute volume departures from reference plane}
\usage{
vol_faceweighted(mesh, radius = 7)
}
\arguments{
\item{mesh}{Object of class "mesh3d"}

\item{radius}{Radius in cm that the mesh was trimmed, see
\code{\link[=pre_process_mesh]{pre_process_mesh()}} and \code{\link[=remove_cyl]{remove_cyl()}}}
}
\value{
List of length 3; each element is a numeric vector of length 1
containing the computed volume above or below the reference datum, or the
sum of these two (absolute value, i.e. total displacement)
}
\description{
Computes area of each face and average distance of its 3
vertices. Multiplies these values to yield a weighted "approximate volume"
for each face. Sums the values based on whether the mean distance is above
or below the reference # datum.
}
\details{
\lifecycle{experimental}
}
\examples{
library(magrittr)
library(soilmesh)
library(mesheR)

trimmed <- remove_cyl(untrimmed_mesh1) \%>\%
  adjust_xy() \%>\%
  adjust_z()


rgl::clear3d()
mesheR::wiremesh3d(trimmed)
add_origin_axes()
mesheR::wiremesh3d(ref_circ)

vols <- vol_faceweighted(trimmed, radius = 7)

}
